Added core18 snap path to PATH.
This should fix the issue where the configure hook doesn't run
properly on systems without netcat installed.

Also got rid of call to install netcat in tools/install_and_build.sh.

Doubled timeout in .zuul.yaml to deal with timeout that happened
during this test (I think that we were pushing the hour timeout as it was).

Change-Id: Ie9cd729109a871c32c12108163d22008d91155f9