Remove iptables part
This change prevents builds in the gate from timing out while pulling
the iptables part, which was happening with increasing frequency.

The part was originally added in an attempt to isolate ourselves
better from the host system, despite being classically confined. This
approach was unreliable -- it doesn't work particularly well for
commands requiring sudo. We can back off this change (especially since
we've removed the special pathing that pointed us at the snapcraft
built iptables anyway -- I think that the part is superfluous).

In addition to the automated tests, I have verified that the following
work:

1. xenial and eoan installs
2. clustering
3. refresh from beta and edge

Change-Id: I7d314eccc0433e54a800ed47add684c2e0fa6565
Closes-bug: 1854081