Fixes for snapcraft 3.9.x
Added build-environment to qemu (works around issue where patchelf was
not finding libs).

Added hack to force wrapping of all non bash apps, to work around
snapd issue with multiple instances of PATH in the snap environment.

Also snuck in fix for horizon build. Horizon is specified in upstream
constraints now, which means that our build fails if we try to build
it from source in the same part that looks at the constriants file.

Misc fixes to make the tests nicer.

Change-Id: I50c88878c4f9dbb07006cab899a717e334be07d0