Move back to using upstream upper-constraints
neutron has been added to the upper constraints file, which means we
can't install it from a tarball (pip gets upset when you try to
install from source while also attempting to specify a version in your
constraints). This commit replaces the reference to the tarball in the
openstack-projects part w/ a reference to neutron from pip.

We still install other OpenStack packages in the part from tarballs,
as the versions of those packages are not locked down in the
upper-constraints.

Change-Id: Id73fbc927c0b56ae6242f2abcaed7b6208f69a9e