Fix pyScss incompatibility w/ latest setuptools
pyScss, which is a build dependency for horizon, currently relies on
the depcrecated Feature class of setuptools. The latest update to
setuptools removed the class entirely, breaking pyScss.

For now, we point to a fork of pyScss 1.3.4, which removes the
reference to the Feature class. In the near future, the upstream
OpenStack project will probably release a better fix; this gets us
unblocked for now.

We point to our custom pyScss by downloading and patching the stein
upper-constraints in our patch app. This is nicer than keeping a local
version of the file in our repo, as we'll automatically get updates,
and our patch will become a no-op when it is no longer needed.

Change-Id: I664ce6402f69d24fa9730c99315cbf65859b0aa3
Closes-bug: 1867360